# openapi.yaml
openapi: 3.1.0
...
info:
  title: Signal Zero API
  version: 2.0.0
  summary: Ranks Nepali settlements by anomalously long SILENCE, not by report volume.
  description: |
    Signal Zero ingests reports about a disaster, resolves them to settlements,
    deduplicates them without a model, and then ranks settlements by how
    anomalously long they have gone with **no confirming report** — an
    exponential time-between-events baseline per hazard-tier/population cohort,
    with a Getis-Ord Gi* neighbourhood statistic for spatial context.

    Four rules constrain the whole surface, and they are visible in the schemas:

    1. **No dispatch.** No field anywhere in this API names a place anyone should
       go or a team anyone should send. Approving a checkpoint item yields a
       non-preferential, alphabetically ordered *shortlist of jurisdictions to
       inform*, never an assignment. Attempting to store a dispatch-shaped key in
       checkpoint evidence raises `DISPATCH_FIELD_FORBIDDEN`.
    2. **Nothing is actionable without a named human.** `POST /api/checkpoint/{id}/approve`
       and `.../reject` return **400 `APPROVER_REQUIRED`** when `approvedBy` is
       missing, not a string, or blank/whitespace-only.
    3. **Zero LLM in dedup scoring or ranking math.** Triage tier 3 is the only
       model touchpoint; its execution counts appear under `stats.tier3`.
    4. **Honest unknowns.** `coverageBasis: "cohort-cold-start"` and a null
...
       silent".

...

...

...

...
    one.

...

...

...

...

...

...
servers:
...
security: []
...
paths:
...
  /problems/{slug}:
    get:
      tags: [problems]
      operationId: getProblemType
      summary: What one problem `type` means.
      parameters:
        - name: slug
          in: path
          required: true
          description: The last segment of a `type` URI, e.g. `approver-required`.
          schema:
            type: string
            examples: [approver-required, invalid-page-size, idempotency-key-reuse]
      responses:
        '200':
          description: The registry entry.
          content:
            application/json:
              schema:
                type: object
                required: [ok, type, status, title, description]
                properties:
                  ok: { type: boolean, const: true }
                  type: { type: string }
                  status: { type: integer }
                  title: { type: string }
                  code: { type: [string, 'null'] }
                  description:
                    type: string
                    description: What this failure means and what the caller can do about it.
                  specification:
                    type: string
                    const: https://www.rfc-editor.org/rfc/rfc9457.html
        '404':
          description: No such problem type is registered.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
...
  schemas:
...
    Problem:
      type: object
      title: Problem
      description: |
        **RFC 9457 Problem Details** — https://www.rfc-editor.org/rfc/rfc9457.html,
        which **obsoletes RFC 7807**. Served as `application/problem+json`.

        THE ONE ERROR SHAPE IN THIS API. Before 2.0.0 there were three
        (`{ok,error}`, `{ok,error,code}`, and a `POST /api/run` 409 carrying
        `reportCount` and `durationMs`), so a client that wanted to branch on a
        failure had only a human-readable English sentence to match on — which made
        every error message load-bearing prose that could not be reworded without
        breaking somebody.

        **Branch on `type`.** It is the stable machine-readable identity. `title`
        and `detail` are prose and may change. Every `type` here dereferences: `GET
        /problems` lists them, `GET /problems/{slug}` explains one.

        `type` values are **relative URI references**, resolved by the client
        against the server origin. Signal Zero has no public domain to hang
        absolute URIs off, and inventing one would produce a URI that resolves to
        nothing on purpose.

        Extension members are permitted by RFC 9457 §3.2 and several problems carry
        them (`runId`, `maximum`, `requested`, `idempotencyKey`, …). They are
        documented on the responses that emit them.
      required: [type, title, status, detail]
      # RFC 9457 section 3.2: a problem document MAY carry extension members, and
      # several here do - `runId` on a run conflict, `requested`/`maximum` on a page
      # size, `idempotencyKey` on a key reuse. A closed schema would make the spec
      # contradict the RFC it claims to implement. The members that matter are
      # documented on the responses that emit them.
      additionalProperties: true
      unevaluatedProperties: true
      properties:
        type:
          type: string
          format: uri-reference
          description: |
            Stable identity of the failure. Dereferences to documentation at the
            same path.
          examples:
            - /problems/approver-required
            - /problems/pipeline-run-in-flight
            - /problems/invalid-page-size
            - /problems/idempotency-key-reuse
        title:
          type: string
          description: Short, human-readable summary of the problem TYPE. Does not change per occurrence.
          examples: ['A named human approver is required']
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: |
            The HTTP status code, duplicated for the benefit of clients that have
            lost it. RFC 9457 §3.1.2: it is advisory and **must agree** with the
            actual response status. A document that disagrees with its own response
            is worse than no document, because a client will believe one of them.
          examples: [400, 404, 409]
        detail:
          type: string
          description: |
            Human-readable explanation specific to THIS occurrence. Never a
            substitute for a value the server does not know — a problem document
            reports what failed, it does not guess at what the caller meant.
        instance:
          type: string
          format: uri-reference
          description: |
            The request URL this occurred on. Carries no personal data, because no
            route in this API takes any.
          examples: ['/api/checkpoint/esc-mfa2j8k-0/approve']
        code:
          type: string
          description: |
            EXTENSION MEMBER, kept for one specific reason: the checkpoint errors
            already carried these codes, thrown from `src/pipeline/checkpoint.js`
            and asserted on by the eval suite at the unit level. Dropping them would
            have forced a rename inside the pipeline to satisfy a change to the HTTP
            layer. They are carried through, not invented. `type` is the identity to
            prefer.
          examples: [APPROVER_REQUIRED, ALREADY_DECIDED, RUN_IN_FLIGHT, INVALID_PAGE_SIZE]
      examples:
        - type: /problems/route-not-found
          title: No such API route
          status: 404
          detail: No route GET /api/settlements
          instance: /api/settlements
          code: NO_ROUTE
          method: GET
          path: /api/settlements

    # -----------------------------------------------------------------------
    # Requests
    # -----------------------------------------------------------------------
...